fix: add constants HTTP_STATUS into separate file httpStatus.js
