+++ backend/controllers/authControllers.js
@@ -1,5 +1,5 @@
 import authServices from '../services/authServices.js';
-import { HTTP_STATUS } from '../constants/auth.js';
+import { HTTP_STATUS } from '../constants/httpStatus.js';
 import ctrlWrapper from '../decorators/ctrlWrapper.js';
 
 const signUpController = async (req, res) => {
