+++ backend/helpers/HttpError.js
@@ -1,12 +1,6 @@
-const messageList = {
-  400: 'Bad Request',
-  401: 'Unauthorized',
-  403: 'Forbidden',
-  404: 'Not Found',
-  409: 'Conflict',
-};
+import { STATUS_MESSAGES } from '../constants/httpStatus.js';
 
-const HttpError = (status, message = messageList[status]) => {
+const HttpError = (status, message = STATUS_MESSAGES[status]) => {
   const error = new Error(message);
   error.status = status;
   return error;
